requirements.txt
Django==2.0.7
django-extensions==2.1.0
djangorestframework==3.8.2
django-webtest==1.9.3
